Adding download sequence script and reorg blast functions
