+++ phylogenetics/formats.py
@@ -13,12 +13,6 @@ from phylogenetics.homologs import Homolog, HomologSet
 # XML parser import
 from xml.etree import ElementTree as ET
 
-# ----------------------------------------------------
-# Defaults BLAST XML parsing tags.
-# ----------------------------------------------------
-
-DEFAULTS = ("Hit_id", "Hit_def", "Hit_len","Hit_accession",
-            "Hsp_hseq", "Hsp_evalue")
 
 # ----------------------------------------------------
 # Methods for parsing different file formats
@@ -68,151 +62,3 @@ def fasta2phylip(lines):
     final_out = "%i  %i\n\n%s\n" % (num_seq,num_columns-10,to_write)
 
     return final_out
-
-
-# ---------------------------------------------------
-# BLAST XML/fasta format
-# ---------------------------------------------------
-
-def flatten_concatenated_XML(xml_string,key_tag):
-    """
-        Clean up naively concatenated XML files by deleting begin/end tags that
-        occur at the place where the two files were concatenated.
-        NOTE: This will break and break royally if the key_tags are on the same
-        lines as other important entries.
-    """
-    input = xml_string.split("\n")
-    set_start = re.compile("<%s>" % key_tag)
-    set_end =   re.compile("</%s>" % key_tag)
-
-    # Find all beginning and end tags...
-    starts = [i for i, l in enumerate(input) if set_start.search(l) != None]
-
-    # If this tag occurs more than once...
-    if (len(starts) != 1):
-
-        # Keep the first start reverse so we are chewing from the bottom.
-        starts.pop(0)
-        starts.reverse()
-
-        # Remove all lines between each end and start, again chewing from the
-        # bottom.
-        for i in range(len(starts)):
-            e = starts[i]
-            while set_end.search(input[e]) == None:
-                input.pop(e),
-                e = e - 1
-            input.pop(e)
-
-    # Return freshly minted, clean XML
-    return "".join(input)
-
-
-def parse_blast_XML(xml_string, tag_list=DEFAULTS):
-    """
-        Parse XML file of hits returned after BLASTing a sequence
-        against a database.
-
-        Args:
-        ----
-        filename: str
-            XML filename returned from BLAST
-        tag_list: tuple
-            Tuple of XML tags that will be included in output data for
-            each sequence.
-
-        Returns:
-        -------
-        all_hits: list of dicts
-            List of sequence data for all hits in XML file (with key,values
-            given by tag_list).
-    """
-
-    # Fix screwed up XML if blasts were done in series...
-    blast_input = flatten_concatenated_XML(xml_string,"BlastOutput_iterations")
-
-    # Read blast file properties (in tag_list) into a list to dump out
-    blast_input = ET.XML(blast_input)
-
-    all_hits = []
-
-    # Navigate to proper level in XML to grab hits
-    for blast_output in blast_input:
-        if blast_output.tag != "BlastOutput_iterations":
-            continue
-
-        for iteration in blast_output:
-            if iteration.tag != "Iteration":
-                continue
-
-            for hits in iteration:
-                # Catch the parent id
-                if hits.tag == "Iteration_query-ID":
-                    parent_id = str(hits.text).strip()
-
-                if hits.tag != "Iteration_hits":
-                    continue
-
-                for hit in hits:
-
-                    # Separate the tag_list out from hits and hit specs.
-                    Hit_list = [t[4:] for t in tag_list if t[0:3] == "Hit"]
-                    Hsp_list = [t[4:] for t in tag_list if t[0:3] == "Hsp"]
-
-                    # Construct a dictionary of tags stripped from XML
-                    properties = dict([(h.tag[4:],str(h.text).strip()) for h in hit])
-                    subset = dict((k, properties[k]) for k in Hit_list)
-                    all_hits.append(subset)
-
-                    for property in hit:
-
-                        if property.tag == "Hit_hsps":
-
-                            for hsp in property:
-
-                                # Construct a dictionary of tags stripped from XML
-                                hsp_properties = dict([(p.tag[4:],str(p.text).strip())
-                                                       for p in hsp])
-                                subset = dict((k, hsp_properties[k]) for k in Hsp_list)
-
-                                # Append inner Hsp properties to list
-                                all_hits[-1].update(subset)
-                                break
-
-    return all_hits, parent_id
-
-
-def parse_blast_fasta(xml_string):
-    """
-        Parse Blast's Fasta/XML formatted file, returning a list of each
-        sequence's data.
-
-        Args:
-        ----------
-        xml_string: str
-            Fasta/XML formatted string from Blast Output.
-
-        Returns:
-        -------
-        sequences: list of dicts
-            List of sequences data in their own lists.
-    """
-
-    # Fix screwed up XML because sequences downloaded and output concatenated
-    sequence_input = flatten_concatenated_XML(xml_string, "TSeqSet")
-
-    # Now we should have valid XML...
-    sequence_input = ET.XML(sequence_input)
-
-    # XML Tag prefix to strip
-    prefix = "TSeq_"
-
-    sequences = []
-    for i, sequence in enumerate(sequence_input):
-        # Rip out all properties of a sequence
-        properties = dict([(s.tag[len(prefix):],str(s.text).strip()) for s in sequence])
-
-        # Append to sequences.
-        sequences.append(properties)
-
-    return sequences
